backend/requirements.txt
dj-database-url==0.5.0
djangorestframework==3.12.1
Django==3.0.8
whitenoise==5.1.0
psycopg2==2.8.6
pre-commit==2.11.1
pytest==6.2.2
django-allauth==0.44.0
dj-rest-auth==2.1.4
python-dotenv==0.17.0
uritemplate==3.0.1
PyYAML==5.4.1
